requirements.txt
Flask==3.0.0
line-bot-sdk
python-dotenv==1.0.0
gunicorn==21.2.0
requests
httpx==0.27.0
anthropic==0.40.0
openai==1.54.3
psycopg2-binary==2.9.9
